[CoreBundle] Adding step for testing badge adding and deleting + adding step + update of adding page

+++ Entity/Badge/Badge.php
@@ -693,7 +693,7 @@ class Badge extends Rulable
     public function postRemove()
     {
         $filePath = $this->getAbsolutePath();
-        if (null !== $filePath) {
+        if (null !== $filePath && is_file($filePath)) {
             unlink($filePath);
         }
     }
